requirements.txt
aiohappyeyeballs>=2.6.1,<3.0
aiohttp>=3.11.18,<4.0
aiosignal>=1.3.2,<2.0
attrs>=25.3.0,<26.0
audioop-lts>=0.2.1,<1.0
discord.py>=2.5.2,<3.0
frozenlist>=1.6.0,<2.0
idna>=3.10,<4.0
multidict>=6.4.3,<7.0
propcache>=0.3.1,<1.0
python-dotenv>=1.1.0,<2.0
yarl>=1.20.0,<2.0
